services/libs/python/mf-nacos-service-registrar/pyproject.toml
[project]
name = "mf-nacos-service-registrar"
version = "0.1.0"
description = "Python Nacos 服务注册/发现二方包，适用于 MF 生态"
authors = [
    {name = "Your Name", email = "your@email.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nacos-sdk-python==2.0.4",
    "aiohttp==3.11.18",
    "pydantic==2.11.3"
]

[tool.setuptools.packages.find]
where = ["src"] 